+++ CHANGELOG.md
@@ -122,6 +122,7 @@ Note that this project **does not** adhere to [Semantic Versioning](https://semv
 - We moved the Git settings into their own preferences tab. [#12630](https://github.com/JabRef/jabref/issues/12630)
 - "Get fulltext", groups "Attach file", "Attach file from URL", "Open folder(s)" and "Open file" commands in right click menu have been moved into a "More file operations..." submenu. [#16829](https://github.com/JabRef/jabref/pull/16829)
 - We changed the logging during full-text search indexing to identify which linked files cause errors. [#15680](https://github.com/JabRef/jabref/issues/15680)
+- We changed autosave into synchronizing: external changes to the library file are merged automatically unless they conflict with unsaved edits. [#8431](https://github.com/JabRef/jabref/issues/8431)
 
 ### Fixed